pyproject.toml
[project]
name = "structure"
version = "0.1.0"
description = ""
authors = [
    {name = "Anna Polyakova",email = "sova408@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "httpx[http2] (>=0.28.1,<0.29.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.0.1,<8.0.0)",
    "pydantic (>=2.4.1,<2.12)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "orjson (>=3.11.4,<4.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "types-redis (>=4.6.0.20241004,<5.0.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pytest-mypy-plugins (>=3.2.0,<4.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "pre-commit (>=4.4.0,<5.0.0)",
    "ruff (>=0.14.5,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "bandit (>=1.8.6,<2.0.0)",
    "pip-audit (>=2.9.0,<3.0.0)",
    "fakeredis (>=2.32.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
acme = "src.app.__main__:app"


# --- Ruff (линтер + часть форматирования) ---
[tool.ruff]
target-version = "py313"
line-length = 79
src = ["src", "tests"]
lint.select = ["E","F","W","I","UP","B","C4","PIE","SIM","PT","PL","Q","RUF"]
lint.ignore = ["E501", "UP032", "B008"]
lint.unfixable = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["acme_app"]
combine-as-imports = true

# --- Black (стабильное форматирование) ---
[tool.black]
line-length = 79
skip-string-normalization = true
target-version = ["py313"]

# --- Mypy (строгий режим с послаблениями на старте) ---
[tool.mypy]
python_version = "3.12"
plugins = []
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
exclude = ["build/", "dist/"]
mypy_path = ["src", "tests"]
# поэтапно разрешай ошибки с помощью per-module конфигов:
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
